# yukka-ui/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring Security OpenApi",
    "description": "Dokumentacja OpenApi dla Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Łukasz Domin",
      "email": "lukaszdomin@interia.pl"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/zmiana-hasla": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HasloRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/zmiana-hasla-email/{email}": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "sendResetPasswordEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/zmiana-email": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "changeEmail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/aktywacja-konta": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/aktywacja-konta/resend": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "confirmResend",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegistrationRequest": {
        "required": [
          "email",
          "haslo",
          "nazwa",
          "powtorzHaslo"
        ],
        "type": "object",
        "properties": {
          "nazwa": {
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_ąćęłńóśźżĄĆĘŁŃÓŚŹŻ]+( [a-zA-Z0-9_ąćęłńóśźżĄĆĘŁŃÓŚŹŻ]+)*$",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "haslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          },
          "powtorzHaslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
      "HasloRequest": {
        "required": [
          "noweHaslo",
          "nowePowtorzHaslo",
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "noweHaslo": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          },
          "nowePowtorzHaslo": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthRequest": {
        "required": [
          "email",
          "haslo"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "haslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Authorization Bearer JWT",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}